requirements.txt
# For all scripts
anthropic
python-dotenv
# For 02_tool_agent.py and 03_protocol_agent.py
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
# For 03_protocol_agent.py and gdocs_mcp_server.py
mcp[cli]